test-knowledge/requirements.txt
Django==4.2.10
django-jet-reboot==1.3.3
django-mptt==0.15.0
djangorestframework
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.1.0
celery==5.3.6
redis==5.0.1 